Additional Shaders + Getting used to Shader Graph
I have made additional progress with the book, I have create couple of new shaders. I feel like I'm starting to understand how shaders are made and I understand more of the code.

On another note, I have also experimented a little with Unity's new shader graph tool. I'm trying to recreate the rain effect in a shader, but its proving to be difficult as the nodes in Unity are a little different to the Unreal ones.