Progress on Shader Graph & Small Steps on CG shader
Graph: I have mode some progress on the Shader Graph version of the Rainy Surface, but for some reason the fade part of the graph doesn't work properly. It is slower compared to the fade-in/out of the water ripples. I wil need to ask my lecturer about this issue, as I've remade this in Unreal and it has worked correctly.

CG: I have started working on the Rainy Surface shader as I know the components required to make it work. The first part is creating a simple Alpha Erosion and combining it with a fade effect to get the ripples to fade-in/out on the surface. I have learned how to use only the red channel of a texture, and I'm able to display it on the surface.