CG_Shaders Progress - Ripple Effect and Out-of-sync fade
With my CG Rainy Surface shader I have added the riple effect using edge mask, the ripples now actually look like ripples that would appear on a surface.

I have added a fade effect to the ripples, so they don't just suddenly disappear but actually fade out and fade it. Just like with the Shader Graph the fade is out-of-sync with the ripples appearing.