Updated Shader Graph, progress on out of sync.
For my Cg shader I have made some progress on getting rid of the out-of-sync problem that I've been having. To work around that I have done the time calculation and stored it into a variable. I have then used that variable accross, and that has seemed to somewhat fixed the issue.

Shadergraph - same progress on the out-of-sync, but now when I lerp both of the ripples effects together the lerps seems to be out of sync (the interpolation calculation). This is the next step I will work on.