Improved the Shader Graph version of the shader & Cg needs a fix.
Shader Graph version has an improved version of the shader with lerping, but the lerping seems to be out-of-sync. The values need to be re-tweaked again to make them match up.
---
Cg version now has the lerping implemented, but I need to add an offset to the uv just like in the shader graph version of the shader to see the difference. At the moment the Cg version looks like it did with the fade.