Updated Lerp, and fixed some timing issue.
One of the issues why the Cg version of the shader was out of sync was because I was using 'half' to store the time, which caused it to go completly out of sync. I now store the times in floats and the effect is much better. I've also began working on the normals and adding colours.