Ripples are finished (just need minor tweaking)
Shader graph version of ripples is now done.

Users can now input a base texture, this is the texture that the ripples will be overlayed onto. Users can add a normal texture that supports the base texture to make the effect more appealing and make the light reflect better.

I have changed the way the normals are handled. They are now a blend between the base texture's normal and the ripples normals.

Its been quite a difficult task making the first part of this shader, as there were many differences between Unity and Unreal, but I had to learn a lot of new things to be able to create the shaders.

More in the upcoming blog.